- Refactoring of variable and treeref expansion into one single pt_expand function   - Renamed the top-level function from pt_expand1() to pt_expand   - Removed CO_FLAGS_REFDONE - size_t return in cbuf_len()

+++ cligen_buf.h
@@ -64,8 +64,8 @@ cbuf    *cbuf_new_alloc(size_t sz);
 
 void     cbuf_free(cbuf *cb);
 char    *cbuf_get(cbuf *cb);
-int      cbuf_len(cbuf *cb); /* XXX size_t */
-int      cbuf_buflen(cbuf *cb);
+size_t   cbuf_len(cbuf *cb); /* XXX size_t */
+size_t   cbuf_buflen(cbuf *cb);
 #if defined(__GNUC__) && __GNUC__ >= 3
 int      cprintf(cbuf *cb, const char *format, ...) __attribute__ ((format (printf, 2, 3)));
 #else
